requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# ML/AI dependencies
openai-whisper>=20231117
keybert>=0.8.2
sentence-transformers>=2.2.2
scikit-learn>=1.3.2
numpy>=1.24.3

# Data processing
pandas>=2.1.4

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development
jupyter>=1.0.0
notebook>=7.0.6

# Whisper and audio processing
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0

# Additional dependencies for Python 3.12 compatibility
setuptools>=68.0.0
wheel>=0.40.0

# Additional dependencies for FastAPI
python-multipart